requirements.txt
Flask==3.1.0
Werkzeug==3.1.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==1.3.24
Flask-Cors==5.0.1
gunicorn==20.1.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
